requirements.txt
fastapi
uvicorn
requests
python-dotenv
pytest
httpx
pytest-cov
flake8
black
isort